pyproject.toml
[project]
name = "retrieve-tailor-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "anthropic>=0.79.0",
  "beautifulsoup4>=4.14.3",
  "pymupdf>=1.26.7",
  "python-dotenv>=1.2.1",
  "requests>=2.32.5"
]

[project.scripts]
retrieve-tailor-example = "retrieve_tailor_example.cli:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
